Reorganized the functions in the Display module by creating Locations and Periods modules
