+++ lib/mehr_schulferien_web/controllers/city_controller.ex
@@ -1,28 +1,22 @@
 defmodule MehrSchulferienWeb.CityController do
   use MehrSchulferienWeb, :controller
 
-  alias MehrSchulferien.{Calendars, Calendars.DateHelpers, Display}
-
-  # def index(conn, _params) do
-  #   federal_states = Display.list_federal_states()
-  #   render(conn, "index.html", federal_states: federal_states)
-  # end
+  alias MehrSchulferien.{Calendars, Calendars.DateHelpers, Periods, Locations}
 
   def show(conn, %{
         "country_slug" => country_slug,
         "city_slug" => city_slug
       }) do
-    location = Display.get_city_by_slug!(city_slug, country_slug)
+    location = Locations.get_city_by_slug!(city_slug, country_slug)
     today = Date.utc_today()
     current_year = today.year
     location_ids = Calendars.recursive_location_ids(location)
-
-    next_12_months_periods = Display.get_12_months_periods(location_ids, today)
+    next_12_months_periods = Periods.get_12_months_periods(location_ids, today)
 
     {next_3_years_headers, next_3_years_periods} =
-      Display.get_3_years_periods(location_ids, current_year)
+      Periods.get_3_years_periods(location_ids, current_year)
 
-    public_periods = Display.get_3_years_public_periods(location_ids, current_year)
+    public_periods = Periods.get_3_years_public_periods(location_ids, current_year)
 
     public_periods =
       Enum.filter(public_periods, &(&1.holiday_or_vacation_type.name != "Wochenende"))
