+++ lib/mehr_schulferien_web/controllers/county_controller.ex
@@ -1,25 +1,26 @@
 defmodule MehrSchulferienWeb.CountyController do
   use MehrSchulferienWeb, :controller
 
-  alias MehrSchulferien.{Calendars, Calendars.DateHelpers, Display, Maps}
+  alias MehrSchulferien.{Calendars.DateHelpers, Periods, Locations}
 
   def show(conn, %{
         "country_slug" => country_slug,
         "federal_state_slug" => federal_state_slug,
         "county_slug" => county_slug
       }) do
-    county = Display.get_county_by_slug!(county_slug, federal_state_slug, country_slug)
-    cities = Maps.list_cities(county)
+    country = Locations.get_country_by_slug!(country_slug)
+    federal_state = Locations.get_federal_state_by_slug!(federal_state_slug, country)
+    county = Locations.get_county_by_slug!(county_slug, federal_state)
+    location_ids = [country.id, federal_state.id, county.id]
+    cities = Locations.list_cities(county)
     today = Date.utc_today()
     current_year = today.year
-    location_ids = Calendars.recursive_location_ids(county)
-
-    next_12_months_periods = Display.get_12_months_periods(location_ids, today)
+    next_12_months_periods = Periods.get_12_months_periods(location_ids, today)
 
     {next_3_years_headers, next_3_years_periods} =
-      Display.get_3_years_periods(location_ids, current_year)
+      Periods.get_3_years_periods(location_ids, current_year)
 
-    public_periods = Display.get_3_years_public_periods(location_ids, current_year)
+    public_periods = Periods.get_3_years_public_periods(location_ids, current_year)
 
     public_periods =
       Enum.filter(public_periods, &(&1.holiday_or_vacation_type.name != "Wochenende"))
