+++ lib/mehr_schulferien_web/controllers/federal_state_controller.ex
@@ -1,7 +1,7 @@
 defmodule MehrSchulferienWeb.FederalStateController do
   use MehrSchulferienWeb, :controller
 
-  alias MehrSchulferien.{Calendars, Calendars.DateHelpers, Display}
+  alias MehrSchulferien.{Calendars.DateHelpers, Periods, Locations}
 
   @digits ["1", "2", "3", "4", "5", "6", "7", "8", "9"]
 
@@ -10,22 +10,22 @@ defmodule MehrSchulferienWeb.FederalStateController do
         "federal_state_slug" => <<first::binary-size(1)>> <> _ = id
       })
       when first in @digits do
-    location = Display.get_federal_state!(id)
+    location = Locations.get_federal_state!(id)
     redirect(conn, to: Routes.federal_state_path(conn, :show, country_slug, location.slug))
   end
 
   def show(conn, %{"country_slug" => country_slug, "federal_state_slug" => federal_state_slug}) do
-    location = Display.get_federal_state_by_slug!(country_slug, federal_state_slug)
+    country = Locations.get_country_by_slug!(country_slug)
+    federal_state = Locations.get_federal_state_by_slug!(federal_state_slug, country)
+    location_ids = [country.id, federal_state.id]
     today = Date.utc_today()
     current_year = today.year
-    location_ids = Calendars.recursive_location_ids(location)
-
-    next_12_months_periods = Display.get_12_months_periods(location_ids, today)
+    next_12_months_periods = Periods.get_12_months_periods(location_ids, today)
 
     {next_3_years_headers, next_3_years_periods} =
-      Display.get_3_years_periods(location_ids, current_year)
+      Periods.get_3_years_periods(location_ids, current_year)
 
-    public_periods = Display.get_3_years_public_periods(location_ids, current_year)
+    public_periods = Periods.get_3_years_public_periods(location_ids, current_year)
 
     public_periods =
       Enum.filter(public_periods, &(&1.holiday_or_vacation_type.name != "Wochenende"))
@@ -37,7 +37,7 @@ defmodule MehrSchulferienWeb.FederalStateController do
     render(conn, "show.html",
       current_year: current_year,
       days: days,
-      location: location,
+      location: federal_state,
       months: months,
       next_12_months_periods: next_12_months_periods,
       next_3_years_headers: next_3_years_headers,
@@ -48,27 +48,28 @@ defmodule MehrSchulferienWeb.FederalStateController do
   end
 
   def faq(conn, %{"country_slug" => country_slug, "federal_state_slug" => federal_state_slug}) do
-    location = Display.get_federal_state_by_slug!(country_slug, federal_state_slug)
-    location_ids = Calendars.recursive_location_ids(location)
+    country = Locations.get_country_by_slug!(country_slug)
+    federal_state = Locations.get_federal_state_by_slug!(federal_state_slug, country)
+    location_ids = [country.id, federal_state.id]
     today = Date.utc_today()
-    todays_public_holiday_periods = Display.list_public_holiday_periods(location_ids, today)
+    todays_public_holiday_periods = Periods.list_public_holiday_periods(location_ids, today)
 
     tomorrows_public_holiday_periods =
-      Display.list_public_holiday_periods(location_ids, Date.add(today, 1))
+      Periods.list_public_holiday_periods(location_ids, Date.add(today, 1))
 
     day_after_tomorrows_public_holiday_periods =
-      Display.list_public_holiday_periods(location_ids, Date.add(today, 2))
+      Periods.list_public_holiday_periods(location_ids, Date.add(today, 2))
 
-    todays_school_free_periods = Display.list_school_free_periods(location_ids, today)
+    todays_school_free_periods = Periods.list_school_free_periods(location_ids, today)
 
     tomorrows_school_free_periods =
-      Display.list_school_free_periods(location_ids, Date.add(today, 1))
+      Periods.list_school_free_periods(location_ids, Date.add(today, 1))
 
     day_after_tomorrows_school_free_periods =
-      Display.list_school_free_periods(location_ids, Date.add(today, 2))
+      Periods.list_school_free_periods(location_ids, Date.add(today, 2))
 
     render(conn, "faq.html",
-      location: location,
+      location: federal_state,
       today: today,
       todays_public_holiday_periods: todays_public_holiday_periods,
       tomorrows_public_holiday_periods: tomorrows_public_holiday_periods,
