+++ test/mehr_schulferien/maps_test.exs
@@ -22,16 +22,6 @@ defmodule MehrSchulferien.MapsTest do
       assert location.name == location_1.name
     end
 
-    test "list_cities/1 returns a certain county's cities" do
-      county = insert(:county)
-      cities = insert_list(3, :city, %{parent_location_id: county.id})
-      cities_ids = cities |> Enum.sort(&(&1.name >= &2.name)) |> Enum.map(& &1.id)
-      _other_city = insert(:city)
-      cities_1 = Maps.list_cities(county)
-      cities_1_ids = Enum.map(cities_1, & &1.id)
-      assert cities_1_ids == cities_ids
-    end
-
     test "get_location!/1 returns the location with given id" do
       location = insert(:federal_state)
       assert Maps.get_location!(location.id) == location
